Use chunkhash instead of hash
With Webpack 3 we were using the `hash` to produce the filename. The
issue with this is that the hash

> Returns the build hash. If any portion of the build changes, this
changes as well.

We only want to be invalidating caches for the specific entries that
have changed as opposed to the entirety of the build. To this end we are
now using `chunkhash` which

> Returns an entry chunk-specific hash. Each entry defined in the
configuration receives a hash of its own. If any portion of the entry
changes, the hash will change as well. [chunkhash] is more granular than
[hash] by definition.

This will allow us to produce more deterministic builds. Unfortunately,
due to the way that we bundle lodash (by turning functions that we don't
use into noops) the `chunkhash` will most likely be different for lodash
on each build. This is something that we can address in the future, but
this commit makes it so that at least lodash is not invalidating the
cache for the entirety of the build.