Make it easy for us invalidate our cached assets
Coming from the world of Sprockets there was a straightforward way for
us to force asset URLs to change by bumping the assets version.

We would like to have the same luxury with our assets generated by
webpack.

Now that we are using chunkhashes, rather than content hashes for our
entry points these assets are less likely to be invalidated regularly
therefore this functionality is even more crucial when debugging CDN
issues.