Enable integrity hashes in test, development and production
After our previous woes with cdn serving up assets of 0 bytes we want to
make it harder for browsers to end up caching and executing bad assets.

To this effect, we are enabling an integrity hashes[1] for assets in our
manifest. We will need to update the main FL app to understand the new
integrity hashes.

1)
https://developer.mozilla.org/en-US/docs/Web/Security/Subresource_Integrity