Use fast-async babel plugin
We want to start using async/await in our app so it's easier to work
with promises. We can't do that easily at the moment as we rely on a
client side polyfill, because Babel transpiles async/await to generators
which is not supported in IE11.

We've configured the plugin to transpile async/await code to promises,
as we already have a polyfill for promises.