Upgrade node-sass 4.11.0
`node-sass` has offically supported Node V10 since 4.9.0 (which was
released back in April 2018) [1].

However with this release they also bumped the version of `lib-sass` to
3.5.3 which deprecated the use of css imports [2].

Due to our extensive use of css imports (within the rails app) we have
been blocked on a potential upgrade since then.

This has been quite a contentious issue [3] and they have since gone back on
their decision to introduce these deprecation warnings.

With the release of node-sass 4.11.0 (November 2018) they have bumped
lib-sass to 3.5.5 which includes the removal of the noisy deprecations
warnings [4].

[1] https://github.com/sass/node-sass/releases/tag/v4.9.0
[2] https://github.com/sass/libsass/releases/tag/3.5.3
[3] https://github.com/sass/node-sass/issues/2362
[4] https://github.com/sass/node-sass/releases/tag/v4.11.0