Allow entries to be a glob
We want to remove the need in the main FL application for a pack to
exist in order to render a component in the view.

To deal with this, we want to make it so that every `index.jsx` file
for a component is now an entry.

This commit makes it so that the entry paths provided in the config
now specify a glob, a root path (which we subtract when producing the
entry name) and a `useDirName` option which determines whether the name
should either use the file name or the directory name.