Simplify NamedChunksPlugin
In
https://github.com/Futurelearn/futurelearn/commit/34529082a0362acade3c27d514b007237b8a5c47
we introduced `NamedChunksPlugin` to ensure chunk and module names are
deterministic in webpack builds.

This was to allow chunks to have names rather than numbers for our
chunkss (i.e. form.123.js as opposed to 1.123.js) and therefore ensure
consistency across builds.

The reason that we added our own logic for it as opposed to the default
logic was to handle the fact that NamedChunksPlugin can only work on
chunks that have a name. Prior to Webpack 2.4, this meant that it
couldn't be used for async chunks (i.e. dynamic imports).

However, given that we currently specify `webpackChunkName` for each of
our dynamic imports, we no longer need this custom logic (as long as
we continue to specify the `webpackChunkName`).

We need to do this, since there's issues with the current logic and
using dynamic imports that don't specify a specific import but rather a
pattern. I've tried going down this rabbit hole but I can't see an end
in sight and I think the risk of someone introducing a dynamic import
(we very rarely do this) without a `webpackChunkName` is worth the time
spent getting to the bottom of this.