Target ECMA 5 for terser minification
When we initially introduced our webpack config, we went with the
default minification options that we were previously using. This means
that when minifying our Webpack assets we are targeting ecma8 (even
though we are using babel to target ecma5).

This has never previously caused us any issues, however we recently
introduced atlassian/react-beautiful-dnd to handle drag and drop. When
producing a minified production build this leads to an error in IE11 of
`ie11 expected hexadecimal digit`.

I believe this is due to the fact that because we are specifying ecma8
as the target, terser is using `PropertyShorthand`s when producing the
minified bundle. IE11 doesn't support this so everything goes boom.

To deal with this, we are now targeting ecma5 as the output which means
terser will no longer attempt to use property shorthands.

Ideally going forward we should tie this version to the ECMA version of
JS that we are targeting.