Use explicit public path, path and hypernova public path
Now that we are dynamically importing most of our client side JS in the
main FL app we can no longer rely on Rails script tags to include the
cdn as part of the src attribute.

Due to this, we now need to be able to explicitly specify the path,
publicPath and hypernovaPublicPath in our webpack.yml so that our
dynamic packs can be served off of cdn in production.

This commit makes it so that our config now uses these values specified
in `config/webpack.yml` in the main fl app.