Symlink src directory
We want to start producing multiple JS bundles so that we can optimise
what we send over the wire dependant on the browser (i.e. modern
browsers understand ES2015 so we shouldn't be sending the transpiled
ES5).

Thankfully Webpack makes this easy to accomplish.

Unfortunately that was before we decided to remove the need for packs
and move to a world where we dynamically import all our React
components.

Why does this make it harder?

There's no way (as far as I can tell) to tell Webpack to dynamically
import something that it has outputted, it can only dynamically import a
real file on disk (that it will then treat as an extra entry path and
run it through it's various pipelines).

So this means that we're left in a position where we can produce
multiple bundles but have no way of dynamically importing them :sadface:

How can we make this work?

If Webpack needs for there to be a real file on disk for it to
dynamically import, we can symlink `app/javascript` to
`app/javascript-legacy` and then we've made Webpack happy. In this
world, we can configure Webpack to run everthing in `app/javascript`
through the modern pipeline and everything in `app/javascript-legacy`
through it's legacy pipeline.

This kind of sucks but I think it's not as bad as having to have a rake
task handle this or actually copy the directory itself.

To handle the symlink, we are using `fs-extra` which adds a bunch of
extra functionality not found within the `fs` module.

When we produce the entries for Webpack, we are now returning 2
versions, `entries` which is all the entry points within
`app/javascript` and `legacyEntries` which is all the entry points
within `app/javascript-legacy`.