Update default babel config to target modern browsers
We want to start producing 2 bundle of client JavaScript. 1 that targets
browsers that support `type=module` (modern browsers) and 1 that targets
browsers that don't.

This will lead to modern browsers receiving less bytes of JS over the
wire and lead to faster parsing and evaluation.

This commit updates the default babel config to target modern browsers
only (using the `esmodules: true` target).

One of the main changes here is that we are now using `useBuiltIns:
'usage` to handle polyfilling of JS language features.

This means that when babel encounters something that needs to be
polyfilled (it makes this decision based on the target) it includes the
necessary polyfill from `@babel/polyfills`.

For instance, if the browsers doesn't support `Promise` given

```
new Promise();
```

babel will rewrite it to look something like

```
import Promise from 'corejs/promise';

new Promise();
```

This means that we are:
a) only polyfilling what we use (as opposed to the whole JS language)
b) we don't have to worry about polyfilling the language (we still need
to worry about polyfilling browsers apis though)

Since we're targeting modern browsers, I've also removed the
`fast-async` plugin (which rewrites `async await` to use promises) since
all browsers that support `type="modules"` support `async await`.

I've also removed `object-rest-spread` for the same reasons.

To allow us to have multiple babel configs we are now also explicitly
stating which babel config to use within the loader.

We've set `modules` to false which instructs babel to not transpile
`import` statements to `require` statements since Webpack understands
`import` statements and gives it the capabilities to statically analyze
our code.