Create es5 babel config
This is the same as our previous default config with the exception that
our target is now `esmodules: false` (i.e. browsers that don't support
`type=module`)

We have also introduced a new webpack loader that uses this config so
that we can apply it in a future commit to our legacy bundles.