Create node babel config
This updates the targets to be `node` which should allow us to have a
more performant server side bundle. `node: 'current'` means that babel
will target the current version of node that is available on the
machine.

Since our version of node supports `async await` and
`object-rest-spread` I've removed those plugins from this config as
well.