Fix null-loader excluding some React components
Sometimes, we do not want to load JS modules during server-side
rendering. For example, in cases where the JS expects the DOM to exist.

To do this we use a null-loader.

Previously however, the null-loader sometimes excluded our own
components, such as for the filename `AWSAttributes.jsx`.

This change makes the null-loader only look in node_modules for matches
to prevent accidentally matching our own components. It also only looks
for matches at the start of a word, so `tributejs` will match but
`attribute` will not.