Update Webpack assets manifest
In
https://github.com/webdeveric/webpack-assets-manifest/commit/051243ec8963be69ab8565b5998032cc3a020a7a
they fixed an issue with memory consumption.

It turns out that producing the manifest is the most expensive part of
rebuilding our assets and this fix greatly improves this. However they
haven't published a version with these changes yet so we are now using
the master branch which has the changes on them.