Only use split chunks on node modules in development
The main reason for this, is to try to reduce the amount of work that
webpack needs to do when watching files for changes by making the module
dependency trees within webpack simpler.

Given that, locally, there is no advantage to having these small chunks
since we can't take advantage of http/2 we are therefore disabling it
for any non vendor assets.

We're leaving vendor assets in so that we don't need to bundle a copy of
React with every component.