Add flag to allow disabling functionality in the test env
We want to be able to turn off scrolling in the test environment because
it's making feature specs fail sporadically if we try to click on
something while scrolling is happening.

This commit adds a constant that only appears in the test environment,
so we can check to see if we should disable the functionality.

I picked a generic name because we may want to disable other things in
future, and we can just use the same name.