Remove hard-source-webpack-plugin
We have recently noticed some inconsistent asset hashes between
different boxes in production. We believe that this is the result of a
caching issue with the `hard-source-webpack-plugin`[1].

The safest thing to do for the time being is to temporarily remove the
plugin whilst we investigate further. This means that our deployment
times will take a hit (~5 mins to build webpack assets) but that feels
better than not being able to deploy safely :-)

[1] https://github.com/mzgoddard/hard-source-webpack-plugin/issues/454