Setup CSS modules
This commit introduces support for CSS modules for a `.module.scss`
file.

The setup is fairly vanilla with 2 exceptions.

We are using `getLocalIdent` so that we can more granuarly control the
classname that gets produced. For us that means it will be

`nameOfDirectory nameOfStylesheet nameOfClass`.

We also need to customise onlyLocals so that for hypernova and our
legacy bundles we are only outputting the names of the classes rather
than the associated styles as well. This means that hypernova will be
able to understand `<div className={styles.name} />` without exploding.