Use corejs 3 with babel preset env
At some unnamed point in the future babel preset env will start using
core-js 3 by default. To prepare for this we are now explicitly telling
babel to use core-js 3.

The main advantage we get here is that babel is now able to optimise the
polyfill combinations which should lead to a smaller bundle for our end
users.