Move @babel/core and babel-loader to peerDependencies
Since installing storybook in the futurelearn app, we've received peer
dependencies warnings about @babel/core and babel-loader when running
yarn install. While these packages are present in our project, they're
not installed at the top level, so yarn warns us about them.

To fix these warnings we need to explicitly depend on these packages in
the main app[1]. To avoid the main app's version of them drifting from
that required by this config package, this commit moves them to be
peerDependencies here, too. To avoid similar warnings when running `yarn
install` in this project, I've also added them as devDependencies.

This does mean that our babel dependencies are now somewhat confusingly
split across the two projects, but this is the minimal change required
to fix the immediate problem.

Longer term we'd like to think about whether the distinction between
this package and the main app is still useful, and whether the
separation between web dependencies and build dependencies can be
managed some other way, e.g. with yarn workspaces.

[1] https://github.com/Futurelearn/futurelearn/pull/2695