Update null-loader regex
Sometimes, we do not want to load packages during server-side rendering.
For example, in cases where the JS expects the DOM to exist. We want
them to be used only the client-side.

To do this we use a combination of a null-loader (which rejects things
that match a pattern from a pack) and a nodeExternals whitelist (which
bundles certain packages instead of reading them from filesystem). See
[1] for an explanation of how this works.

We use one shared `CLIENT_SIDE_ONLY_PACKAGES` regex for both the
null-loader and the whitelist.

Recently we found some Jenkins CI builds failing if they included in
their branch name `c3`, `d3` etc. [2]

This happens because the nodeExternals whitelist is comparing the regex
to just the package name, but the null-loader is comparing the regex
against the full filepath, which includes the branch name.

Here we update `client_side_only_packages.js` to provide two different
loader tests, one for just names and another for full paths.

[1] https://github.com/Futurelearn/webpack-config/commit/01f2fc63c0eb69213cc7d5848915afd2bc7c47ac
[2] https://futurelearn.slack.com/archives/C025SEBUG/p1565108921303000