Use ASSET_HOST environment variable
We want to be test compiling our webpack assets with an alternative
hostname, without having to hard code it, and without it interfering
with our exsitng `public_path` and `hypernova_public_path`.

This commit switches the `output.publicPath` and
`output.hypernovaPublicPath` to use an `ASSET_HOST` environment
variable if it's present, otherwise it will fall back to the regular
config.

Co-authored-by: Tom Marshall <tom.marshall@futurelearn.com>
Co-authored-by: Laura Martin <laura.martin@futurelearn.com>