Add jQuery as an external
We currently include jQuery in our app via Sprockets. However, some
packages we bring in via Yarn and Webpack also depend on it. Currently,
they load their own version of jQuery and don't use the global instance.

Adding `jquery` as an "external" tells Webpack that whenever a file
attempts to import the module named `jquery`, to not bother loading it,
and use the existing `jQuery` global instead.

[1]: https://webpack.js.org/configuration/externals/#externals