Load safari-nomodule.js as a static file
The `safari-nomodule.js` polyfill needs to be inserted ealy on in the
page before any other script tags with `type="module"` or `nomodule`
attributes. As such, we don't really want to process it with Webpack,
and we don't want to insert it in the page _after_ `runtime.js`.

Using `file-loader`, we'll get an entry in `manifest.json` named
`app/javascript/vendor/safari-nomodule.js` that we can load with a
regular `script` tag wherever/however we want.