Use git+https protocol for webpack-assets-manifest
Whenever we update the FutureLearn application to use the latest version
of our webpack-config package, yarn switches between https and git+https
for getting the webpack-assets-manifest package. We're not sure what's
causing this behaviour, so I'm trying this to see if it stops the
weirdness. ¯\_(ツ)_/¯

(See https://docs.npmjs.com/files/package.json#git-urls-as-dependencies
for more info on the supported protocols.)