Use separate runtime for "init" scripts
We currently use a single runtime which is added to the page just before
the first pack. Because we use the `type="module"` attribute on our
script tags, the loading `runtime.js` (and any other chunks) is
"deferred", meaning it is only executed once the document has been
parsed (but before the `DOMContentLoaded` event) [1].

However, for some JavaScript, we want to load and execute right away as
early on, even before the document has been parsed. In order to do this,
we don't want to use `type="module"` on the script tag for that pack.
That means we can't use a deferred runtime script either.

This commit changes the `runtimeChunk` setting to produce two runtime
chunks: the existing `runtime.js` which we'll continue to use for
deferred packs, and a separate `init.js` that we can use for blocking
packs (in this case there's inly one `packs/applicationInit.js`).

[1]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/script#Attributes