Only use style-loader for CSS Modules
In order to move our stylesheets to webpack we want real CSS files for
the main stylesheets in our app instead of using the style-loader.

This is also necessary as the Premailer gem requires an actual
stylesheet to inline styles in emails.

This commit ensures we use `MiniCssExtractPlugin` for everything except
CSS modules, which will still use the `style-loader`. `style-loader`
will automatically inject styles into the DOM using multiple style tags.
MiniCssExtractPlugin will output CSS files that can be used with link
tags.

Co-authored-by: Tom Marshall <tom.marshall@futurelearn.com>
Co-authored-by: Tara Ojo <tara.ojo@futurelearn.com>