Don't publicly expose font files
We need webpack to process our font files so that they can be used for
our print proof product PDFs; however, we do not want to expose these
files publicly, as we'd be in breach of our license.

This commit updates our config so that the file loader does not process
the font files outside of the bundles created for use on the server.