Process print proof product styles
We need to process the print proof product styles as they'll be used
server-side for PDF generation; however, we want to avoid processing
every other stylesheet unnecessarily.

For our server assets we replace SASS files with empty modules using
null-loader. This commit ensures that we exclude the print proof product
files from the null loader so that they are processed correctly by
sass-loader.