Use separate entrypoints for server/hypernova
This commit requires us to list separate entrypoints for
server/hypernova assets by tagging them with a `server` flag.

This separation will allow us to be more selective about which files we
actually need to compile for server/hypernova use, and likewise stop
compiling some assets that _aren't_ used client-side.

I chose the to name the flag `server`, as it better describes that some
of these assets are for use by hypernova, and others are used for
server-side rendering of PDF certificates.