Allow webpack to generate hard-coded stylesheets in development
Trello: https://trello.com/c/YTi3Gjfz

In development, we inject CSS Modules stylesheets via JavaScript.
This is to keep the build time low, because the CSS extract plugin
is quite slow.

The downside of this is that these stylesheets don't get loaded if JS
is disabled.

However, as we move towards using CSS Modules more widely, and
especially once we start cutting the mustard (ie. not serving JS to
older browsers), it's important that we should be able to test the site
with JS disabled and have it look as it would in production.

This commit updates the webpack config to allow use of a
`USE_STYLESHEETS` environment variable that allows us to tell webpack to
run as if in production (ie. extract JS-imported stylesheets and include
them in a way that is not dependent on JS).

Initially we can use it directly from the command line:

    USE_STYLESHEETS=true ./bin/webpack

...but there might be ways we could make this easier to use.

Co-authored by: wannabefro <sam.mctaggart@gmail.com>