Stop creating symlink for legacy assets
We compile a set of "legacy" assets alongside our regular "modern"
assets, to ensure we serve smaller bundles to newer browsers whilst also
polyfilling behaviour for older browsers.

Previously, we created a symlink of our source files directory and
used that with our legacy config. However, this was slower and more
complicated than necessary, and have figured out an alternative
approach. As such, we no longer need to create this symlink.