Replace server and legacy options with builds
To compile a full set of assets for our app we have a "main" build, a
"legacy" build, and a "server" build. There is some crossover, with some
entry paths being processed by all three builds, whilst others are only
processed by one or two builds.

This resulted in a rather complex YAML config with repeated entry paths
and boolean options and implicit defaults etc.

This commit simplifies the config by removing the `server` and `legacy`
boolean options and the implicit default for the main build, and
replacing them with a `builds` array, which will allow us to get rid of
the repeated entry paths.